Build fails on aarch64
[Issue]    http://web.sec.samsung.net/bugzilla/show_bug.cgi?id=15312
[Problem]  Incompatible types used in 32 and 64 bit architecture
[Solution] Use proper types
[Verify]   Build and run browser in both architectures.

To build aarch64 uncomment proper line in build_mobile.sh

+++ services/QuickAccess/QuickAccess.cpp
@@ -453,7 +453,7 @@ void QuickAccess::addBookmarkManagerTile()
     elm_gengrid_item_append(m_bookmarkGengrid, m_bookmarkManagerTileclass, this, _bookmark_manager_clicked, this);
 }
 
-void QuickAccess::setIndexPage(const int page) const
+void QuickAccess::setIndexPage(const uintptr_t page) const
 {
     Elm_Object_Item* it = elm_index_item_find(m_index, (void *)page);
     if (it != NULL) {
